requirements.txt
pillow==11.1.0
qrcode==8.0